cleanup token contracts
