+++ examples/token/wallet/src/lib.rs
@@ -1,4 +1,4 @@
-use athena_interface::Address;
+use athena_interface::{Address, MethodSelector};
 use parity_scale_codec::{Decode, Encode};
 
 #[derive(Decode, Encode)]
@@ -16,4 +16,18 @@ pub struct SpendArguments {
   pub token_identifier: Address,
   pub recipient: Address,
   pub amount: u64,
+}
+
+pub const SELECTOR_RECEIVE: MethodSelector = MethodSelector([208, 20, 84, 100]);
+
+#[cfg(test)]
+mod tests {
+  use athena_interface::MethodSelector;
+
+  use crate::SELECTOR_RECEIVE;
+
+  #[test]
+  fn receive_method_selector() {
+    assert_eq!(SELECTOR_RECEIVE, MethodSelector::from("athexp_receive"));
+  }
 }
